refactor: Class name changed.
